+++ project/sanghun-onboarding/src/main/java/com/example/survey/controller/SurveyController.java
@@ -1,0 +1,5 @@
+package com.example.survey.controller;
+
+public class SurveyController {
+
+}
